lit: restore UNSUP on missing binaries
some tweaks to wind back my changes to lit.py,
hopefully making it more friendly to the CI build which runs
only alive2 tests.

there's a bit of a hack in build.sh.
when building with LLVM from Nix, the alivecc/alive++ will be non-functional
due to an incorrect clang path.
this would lead to test failures, so if built by nix, we remove the
alivecc/alive++ binaries to mark the clang-tv as unsupported.
these tests should still run correctly on github actions and
if you have a non-nix LLVM.