add metadata annotations with asm name
adds metadata strings prefixed by ! which attach
the mcinst name to the llvm instruction.
the metadata is stored in a node called !asm.aslp
or !asm.classic, depending on how that instruction
was lifted.

note: for the classic lifer, the !asm.classic metadata
is attached to the branch /preceding/ the basic block
of that instruction,

for example, below, the !2 instruction is the lifter_a3_0
basic block. the metadata values are listed at the end of
the LLVM module output.

lifter_a2_0:                                      ; preds = %.Lfunc_begin0
  br label %lifter_a3_0, !asm.classic !2

lifter_a3_0:                                      ; preds = %lifter_a2_0
  %a3_1 = load i128, ptr %Q0, align 1
  %a3_2 = trunc i128 %a3_1 to i64
  %a3_3 = bitcast i64 %a3_2 to <8 x i8>
  %a3_4 = sub <8 x i8> zeroinitializer, %a3_3
  ...

!2 = !{!"NEGv8i8"}